(BSR) fix(cheatcode): remove cheatcode snapshot and reorganise cheatcode components
