+++ src/features/cheatcodes/components/CheatCodesButton/CheatCodesButton.native.test.tsx
@@ -1,16 +1,16 @@
 import React from 'react'
 
 import { navigate } from '__mocks__/@react-navigation/native'
-import { render, fireEvent } from 'tests/utils/web'
+import { render, fireEvent } from 'tests/utils'
 
 import { CheatCodesButton } from './CheatCodesButton'
 
 describe('CheatCodesButton component', () => {
-  it('should have a button to go to the Login', () => {
+  it('should have a button to go to the cheatcodes section', () => {
     const component = render(<CheatCodesButton />)
 
     const Button = component.getByText('CheatCodes')
-    fireEvent.click(Button)
+    fireEvent.press(Button)
 
     expect(navigate).toHaveBeenCalledWith('CheatCodes')
   })
