+++ src/features/cheatcodes/pages/CheatCodes/CheatCodes.tsx
@@ -4,8 +4,8 @@ import { Text, Alert, Button, ScrollView } from 'react-native'
 import styled from 'styled-components/native'
 
 import { api } from 'api/api'
-import { CodePushButton } from 'features/cheatcodes/components/CodePushButton'
-import { CrashTestButton } from 'features/cheatcodes/components/CrashTestButton'
+import { CodePushButton } from 'features/cheatcodes/components/CodePushButton/CodePushButton'
+import { CrashTestButton } from 'features/cheatcodes/components/CrashTestButton/CrashTestButton'
 import { NavigateHomeButton } from 'features/cheatcodes/components/NavigateHomeButton/NavigateHomeButton'
 import { RootStackParamList } from 'features/navigation/RootNavigator/types'
 import { env } from 'libs/environment'
